requirements.txt
asgiref==3.2.3
Django==3.0.3
django-extensions==2.2.7
djangorestframework==3.11.0
mysqlclient==1.4.6
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
Werkzeug==1.0.0
